requirements.txt
fastapi==0.111.0
gunicorn==21.2.0
uvicorn==0.30.1
langgraph 
langchain_openai 
langchain_core
python-dotenv
pyodbc==5.2.0
azure-search-documents
pytz
aioodbc==0.5.0
azure-ai-projects==1.0.0b6
azure-identity
langchain-azure-ai
langchain-azure-ai[opentelemetry]
matplotlib==3.10.3
seaborn==0.13.2
azure-storage-blob==12.25.1
openpyxl==3.1.5
tabulate==0.9.0
aiofiles==24.1.0